new implementation, injecting a Binder service into the phone process
In theory, the technique can work on any device, possibly with some
adjustments for different Android versions or vendor patches.

It hooks the standard telephony service, and responds to
requests that have its interface descriptor.

Due to a bug with the build script, you'll have to build twice to
properly package the dex file that gets injected into the phone
process.  Also, make sure to sync git submodules.

Also added on-screen reporting of the retrieved values.