MTK workaround - omit path
I don't really understand why, but this is reported to work.

I got the idea by comparing AT commands in logs from the
device's telephony code -- those were omitting the path.

But when I disassemble the code, I don't see where the path
is being dropped, so maybe the "true" logic is more complex
than this.