RilExtender: more refactoring and cleanup
Unfortunately, we can get in a bad state if `com.android.phone` goes
away and someone is trying to broadcast to it.  Things hang for
a while (maybe only because I block waiting for a response), then
eventually I get a bunch of stacktraces about Binder objects being
dead, then sometimes com.android.phone dies again, etc.  Probably the
solution would be to properly un-register for the broadcasts, or move
to another IPC channel.